fix(bench): split the real-corpus leg into its own bench target
The documented `cargo bench --bench json_validate_bench -- validate_real_corpus`
still panicked on a fresh checkout: `Criterion::filter_matches` is private, so
a `--` filter cannot stop `criterion_group!` from executing every group
function, including the synthetic-ladder ones that assert the whole
pattern/size matrix is present before registering a single benchmark. Passing
a filter never exempted the real-corpus leg from needing 55 unrelated files.

Moved it into benches/json_validate_real_corpus_bench.rs, a standalone
`[[bench]]` target with no dependency on data/bench/generated/. Verified in a
tree without that directory: the new target exits 0 and reports 3 seed files;
json_validate_bench still fails loudly and correctly when its ladder is
missing, unchanged.
