+++ Cargo.toml
@@ -207,6 +207,10 @@ name = "json_validate_bench"
 harness = false
 
 [[bench]]
+name = "json_validate_real_corpus_bench"
+harness = false
+
+[[bench]]
 name = "utf8_validate_bench"
 harness = false
